# wwwroot/openapi.json
{
"openapi": "3.0.3",
"info": {
  "title": "TripTastic API",
  "description": "Travel API for searching and booking flights and hotels. Supported destinations: New York (JFK), Los Angeles (LAX), Chicago (ORD), Dallas (DFW), Denver (DEN), San Francisco (SFO), Seattle (SEA), Miami (MIA), Boston (BOS), Atlanta (ATL), London (LHR), Paris (CDG), Frankfurt (FRA), Tokyo (NRT), Sydney (SYD). IMPORTANT: Call /api/System/current-date first to get today's date.",
  "version": "1.0.0"
},
"servers": [{"url": "https://triptastic.azurewebsites.net"}],
"paths": {
...
  "/api/Cart/flights": {
    "post": {
      "operationId": "addFlightToCart",
      "summary": "Add a flight to the shopping cart",
      "description": "Adds a flight to the cart. Use the flight ID from planTrip results. Add both outbound and return flights.",
      "requestBody": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/AddFlightToCartRequest"}
          }
        }
      },
      "responses": {
        "200": {"description": "Success", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CartResponse"}}}},
        "400": {"description": "Bad request"},
        "404": {"description": "Flight not found"}
      }
    }
  },
  "/api/Cart/hotels": {
    "post": {
      "operationId": "addHotelToCart",
      "summary": "Add a hotel to the shopping cart",
      "description": "Adds a hotel reservation to the cart. Use the hotel ID from planTrip results.",
      "requestBody": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/AddHotelToCartRequest"}
          }
        }
      },
      "responses": {
        "200": {"description": "Success", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CartResponse"}}}},
        "400": {"description": "Bad request"},
        "404": {"description": "Hotel not found"}
      }
    }
  },
  "/api/Cart/checkout": {
    "post": {
      "operationId": "checkout",
      "summary": "Checkout and book the trip",
      "description": "Completes the checkout process, books all items in the cart, and creates a confirmed trip. The cart is cleared after successful checkout.",
      "responses": {
        "200": {"description": "Success", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CheckoutResponse"}}}},
        "400": {"description": "Cart is empty"}
      }
    }
  },
...
  "/api/Cart/trips/{tripId}": {
    "get": {
      "operationId": "getTrip",
      "summary": "Get a specific booked trip",
      "description": "Returns details of a specific booked trip by ID.",
      "parameters": [
        {"name": "tripId", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}, "description": "The trip ID"}
      ],
      "responses": {
        "200": {"description": "Success", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BookedTripResponse"}}}},
        "404": {"description": "Trip not found"}
      }
    }
  },
  "/api/Cart/trips/{tripId}/cancel": {
    "post": {
      "operationId": "cancelTrip",
      "summary": "Cancel a booked trip",
      "description": "Cancels a previously booked trip.",
      "parameters": [
        {"name": "tripId", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}, "description": "The trip ID to cancel"}
      ],
      "responses": {
        "200": {"description": "Success", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CancelTripResponse"}}}},
        "404": {"description": "Trip not found"}
      }
    }
  }
},
"components": {
"schemas": {
...
    "AddFlightToCartRequest": {
      "type": "object",
      "required": ["flightId"],
      "properties": {
        "flightId": {"type": "string", "format": "uuid", "description": "The flight ID from planTrip results"},
        "passengers": {"type": "integer", "default": 1, "description": "Number of passengers"}
      }
    },
    "AddHotelToCartRequest": {
      "type": "object",
      "required": ["hotelId", "checkInDate", "checkOutDate"],
      "properties": {
        "hotelId": {"type": "string", "format": "uuid", "description": "The hotel ID from planTrip results"},
        "checkInDate": {"type": "string", "format": "date", "description": "Check-in date (same as departure date)"},
        "checkOutDate": {"type": "string", "format": "date", "description": "Check-out date (same as return date)"},
        "rooms": {"type": "integer", "default": 1, "description": "Number of rooms"},
        "guests": {"type": "integer", "default": 1, "description": "Number of guests"}
      }
    },
    "CartResponse": {
      "type": "object",
      "properties": {
        "items": {"type": "array", "items": {"$ref": "#/components/schemas/CartItem"}},
        "totalItems": {"type": "integer"},
        "totalPrice": {"type": "number"}
      }
    },
    "CartItem": {
      "type": "object",
      "properties": {
        "id": {"type": "string", "format": "uuid"},
        "type": {"type": "string", "enum": ["flight", "hotel"]},
        "description": {"type": "string"},
        "totalPrice": {"type": "number"},
        "flight": {"$ref": "#/components/schemas/CartFlightInfo"},
        "hotel": {"$ref": "#/components/schemas/CartHotelInfo"}
      }
    },
    "CartFlightInfo": {
      "type": "object",
      "properties": {
        "id": {"type": "string", "format": "uuid"},
        "airline": {"type": "string"},
        "flightNumber": {"type": "string"},
        "origin": {"type": "string"},
        "destination": {"type": "string"},
        "departureTime": {"type": "string", "format": "date-time"},
        "arrivalTime": {"type": "string", "format": "date-time"},
        "pricePerSeat": {"type": "number"},
        "passengers": {"type": "integer"}
      }
    },
    "CartHotelInfo": {
      "type": "object",
      "properties": {
        "id": {"type": "string", "format": "uuid"},
        "name": {"type": "string"},
        "location": {"type": "string"},
        "starRating": {"type": "integer"},
        "pricePerNight": {"type": "number"},
        "checkInDate": {"type": "string", "format": "date"},
        "checkOutDate": {"type": "string", "format": "date"},
        "nights": {"type": "integer"},
        "rooms": {"type": "integer"},
        "guests": {"type": "integer"}
      }
    },
    "CheckoutResponse": {
      "type": "object",
      "properties": {
        "success": {"type": "boolean"},
        "message": {"type": "string"},
        "bookedTrip": {"$ref": "#/components/schemas/BookedTripResponse"}
      }
    },
...
    "BookedTripResponse": {
      "type": "object",
      "properties": {
        "id": {"type": "string", "format": "uuid"},
        "confirmationCode": {"type": "string"},
        "status": {"type": "string"},
        "bookedAt": {"type": "string", "format": "date-time"},
        "totalPrice": {"type": "number"},
        "flights": {"type": "array", "items": {"$ref": "#/components/schemas/BookedFlightInfo"}},
        "hotels": {"type": "array", "items": {"$ref": "#/components/schemas/BookedHotelInfo"}}
      }
    },
    "BookedFlightInfo": {
      "type": "object",
      "properties": {
        "id": {"type": "string", "format": "uuid"},
        "confirmationCode": {"type": "string"},
        "airline": {"type": "string"},
        "flightNumber": {"type": "string"},
        "origin": {"type": "string"},
        "destination": {"type": "string"},
        "departureTime": {"type": "string", "format": "date-time"},
        "arrivalTime": {"type": "string", "format": "date-time"},
        "passengers": {"type": "integer"},
        "pricePerSeat": {"type": "number"},
        "totalPrice": {"type": "number"}
      }
    },
    "BookedHotelInfo": {
      "type": "object",
      "properties": {
        "id": {"type": "string", "format": "uuid"},
        "confirmationCode": {"type": "string"},
        "hotelName": {"type": "string"},
        "location": {"type": "string"},
        "starRating": {"type": "integer"},
        "checkInDate": {"type": "string", "format": "date"},
        "checkOutDate": {"type": "string", "format": "date"},
        "nights": {"type": "integer"},
        "rooms": {"type": "integer"},
        "guests": {"type": "integer"},
        "pricePerNight": {"type": "number"},
        "totalPrice": {"type": "number"}
      }
    },
    "CancelTripResponse": {
      "type": "object",
      "properties": {
        "success": {"type": "boolean"},
        "message": {"type": "string"},
        "tripId": {"type": "string", "format": "uuid"}
      }
    }
  }
}
}